implement permissions tests
